Move chart data processing functions out of components
