+++ src/components/DataGraph/__tests__/DataGraph-test.js
@@ -1,5 +1,5 @@
 jest.dontMock('../DataGraph');
-jest.dontMock('../util');
+jest.dontMock('../../../core/util');
 
 import React from 'react';
 import TestUtils from 'react-addons-test-utils';
@@ -53,7 +53,7 @@ describe('DataGraph', function() {
   });
 
   it('accepts data', function() {
-    var parseDataForC3 = require('../util').parseDataForC3;
+    var parseDataForC3 = require('../../../core/util').parseDataForC3;
 
     var data = parseDataForC3(timeseriesTestData);
